+++ AsyncEchoServer/AsyncServerDriver/Program.cs
@@ -14,48 +14,13 @@ namespace AsyncServerDriver
         {
             try
             {
-                var server = new AsyncServer();
-
-                server.AsyncServerEvent += AsyncServer_ServerEvent;
-
+                var server = new AsyncServerLib.AsyncEchoServer();
                 server.Start(new IPEndPoint(IPAddress.Loopback, 13000)).Wait();
             }
             catch (Exception ex)
             {
                 Console.WriteLine(ex.ToString());
             }
-        }
-
-        private static void AsyncServer_ServerEvent(object sender, AsyncServerEventArgs e)
-        {
-            switch (e.LastAction)
-            {
-                case LastAction.Accept:
-                    e.NextAction = NextAction.Receive;
-                    e.ClearBuffer();
-                    break;
-                case LastAction.Receive:
-                    if (HasLf(e.Buffer, e.Count) || e.BytesTransferred == 0)
-                    {
-                        e.NextAction = NextAction.Send;
-                    }
-                    else
-                    {
-                        e.NextAction = NextAction.Receive;
-                    }
-                    break;
-                case LastAction.Send:
-                    e.NextAction = NextAction.Shutdown;
-                    break;
-                default:
-                    break;
-            }
-        }
-
-        private static bool HasLf(byte[] buf, int count)
-        {
-            const byte LF = 0xA;
-            return buf.Take(count).FirstOrDefault(b => b == LF) == LF;
         }
     }
 }
