+++ AsyncEchoServer/AsyncServerLib/AsyncServerLib.csproj
@@ -42,6 +42,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="AsyncEchoServer.cs" />
     <Compile Include="AsyncServer.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
